+++ src/pages/dataconsumption/dataconsumption/IDSDataConsumptionPage.js
@@ -28,7 +28,7 @@ export default {
             downloadLink: "",
             dialog: false,
             valid: false,
-            providerUrlRule: validationUtils.getProviderUrlRequiredRule(),
+            providerUrlRule: validationUtils.getUrlRequiredRule(),
             headers: [{
                 text: 'Creation date',
                 value: 'creationDate',
@@ -108,7 +108,7 @@ export default {
             try {
                 this.$data.receivedCatalogs = await dataUtils.receiveCatalogs(this.$data.recipientId);
             } catch (error) {
-                errorUtils.showError(error, "Receive Catalogs");
+                errorUtils.showError(error, "Request Catalogs");
             }
             this.$data.noCatalogsFound = this.$data.receivedCatalogs.length == 0;
             this.$root.$emit('showBusyIndicator', false);
@@ -119,8 +119,17 @@ export default {
             try {
                 this.$data.resourcesInSelectedCatalog = await dataUtils.receiveResourcesInCatalog(this.$data.recipientId, catalogID);
             } catch (error) {
-                errorUtils.showError(error, "Receive Resources in Catalog");
+                errorUtils.showError(error, "Request Resources in Catalog");
             }
+            this.$data.idsResourceCatalog = {};
+            this.$data.selectedResource = {};
+            this.$data.selectedRepresentations = [];
+            this.$data.selectedRepresentation = {};
+            this.$data.selectedArtifacts = [];
+            this.$data.selectedIdsArtifact = {};
+            this.$data.idsContractOffer = {};
+            this.$data.requestContractResponse = {};
+            this.$data.downloadLink = "";
             this.$root.$emit('showBusyIndicator', false);
         },
 
@@ -134,7 +143,7 @@ export default {
             try {
                 this.$data.idsResourceCatalog = await dataUtils.receiveIdsResourceCatalog(this.$data.recipientId, catalogId);
             } catch (error) {
-                errorUtils.showError(error, "Receive Resources in Catalog");
+                errorUtils.showError(error, "Request Resources in Catalog");
             }
         },
 
@@ -150,13 +159,13 @@ export default {
                     try{
                         this.$data.selectedIdsArtifact = await dataUtils.receiveIdsArtifact(this.$data.recipientId, artifact["@id"])
                     } catch (error) {
-                        errorUtils.showError(error, "Receive Artifact");
+                        errorUtils.showError(error, "Request Artifact");
                     }
         
                     try{
                         this.$data.idsContractOffer = await dataUtils.receiveIdsContractOffer(this.$data.recipientId, artifact["@id"])
                     } catch (error) {
-                        errorUtils.showError(error, "Receive Contract Offer");
+                        errorUtils.showError(error, "Request Contract Offer");
                     }
                 }, */
 
@@ -169,8 +178,10 @@ export default {
             try {
                 this.$data.requestContractResponse = await dataUtils.receiveContract(this.$data.recipientId,
                     this.$data.selectedResource["@id"], this.$data.selectedResource["ids:contractOffer"], this.$data.selectedIdsArtifact, download);
+                let agreementId = await dataUtils.getIdOfAgreement(this.$data.requestContractResponse._links.artifacts.href);
+                this.$data.downloadLink = (await dataUtils.getAgreementArtifacts(agreementId))[0]._links.data.href;
             } catch (error) {
-                errorUtils.showError(error, "Receive Contract");
+                errorUtils.showError(error, "Request Contract");
             }
         },
 
@@ -195,7 +206,12 @@ export default {
                     }
                 }
             });
-
+            this.$data.selectedRepresentation = {};
+            this.$data.selectedArtifacts = [];
+            this.$data.selectedIdsArtifact = {};
+            this.$data.idsContractOffer = {};
+            this.$data.requestContractResponse = {};
+            this.$data.downloadLink = "";
         },
 
 
@@ -210,7 +226,6 @@ export default {
 
         clickAcceptContract(artifact) {
             this.$data.selectedIdsArtifact = artifact;
-            this.$data.downloadLink = artifact["@id"] + "/data";
             this.requestContract();
             this.subscribeToResource();
         }