+++ src/pages/dataoffering/routes/addroute/dialog/editidsendpointdialog/EditIdsEndpointDialog.js
@@ -32,7 +32,7 @@ export default {
             let resource;
             if (node == null) {
                 this.$data.title = "Add IDS Endpoint";
-                resource = clientDataModel.createResource(-1);
+                resource = clientDataModel.createResource("", -1);
             } else {
                 this.$data.title = "Edit IDS Endpoint";
                 resource = node.resource;
@@ -41,7 +41,7 @@ export default {
             this.$refs.addResourcePage.set(resource, false);
             this.$data.dialog = true;
         },
-        saved(title, description, language, keywords, version, standardlicense, publisher, policyDescriptions, filetype, bytesize, brokerList) {
+        saved(title, description, language, paymentMethod, keywords, version, standardlicense, publisher, policyDescriptions, filetype, bytesize, brokerList) {
             let isNew = false;
             if (this.$data.node == null) {
                 isNew = true;
@@ -61,6 +61,7 @@ export default {
             resource.title = title;
             resource.description = description;
             resource.language = language;
+            resource.paymentMethod = paymentMethod;
             resource.keywords = keywords;
             resource.version = version;
             resource.standardlicense = standardlicense;